Fake rate making macros
